% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsAddTrees.R
\name{fvsAddTrees}
\alias{fvsAddTrees}
\title{Add Trees}
\usage{
fvsAddTrees(newtrees)
}
\arguments{
\item{newtrees}{A dataframe that has values defined for the following 
attributes of new trees:
`dbh`: Diameter breast height (inches)
`species`: FVS numeric species code
`ht`: Height (feet)
`cratio`: Crown ratio (proportion of height in live crown
`plot`: FVS numeric plot index
`tpa`: Trees per acre}
}
\value{
A scalar integer with the return code from FVS where 0 signals trees 
were added and 1 signals an error.
}
\description{
Add new tree records to an existing simulation. Designed to be 
used to simulate regeneration establishment.
}
