% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsCompositeSum.R
\name{fvsCompositeSum}
\alias{fvsCompositeSum}
\title{Create composite summary}
\usage{
fvsCompositeSum(allsum)
}
\arguments{
\item{allsum}{A list of summary statistics tables as returned by `fvsGetSummary()`.}
}
\value{
A named list of 4 matrices. Each matrix contains entries only for 
the years that are in common to all of the summary tables in sumList. 
The matrices are as follows:
`sumTable`: The weighted average of the individual summary tables, with the same 
column headings
`ForType`: The column headings are the unique forest type codes found in the 
data and the data are the sum of the corresponding sampling weights
`SizeCls`: The column headings are the unique size classes found in the data 
and the data are the sum of the corresponding sampling weights.
`StkCls`: The column headings are the unique stocking classes found in the data 
and the data are the sum of the corresponding sampling weights.
}
\description{
Create composite summary
}
