% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetDims.R
\name{fvsGetDims}
\alias{fvsGetDims}
\title{Get dims}
\usage{
fvsGetDims()
}
\value{
A named vector of integers as follows: 
  `ntrees`, `ncycles`, `nplots`, `maxtrees`, `maxspecies`, `maxplots`, 
  `maxcycles`
}
\description{
Get dims
}
