% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetEventMonitorVariables.R
\name{fvsGetEventMonitorVariables}
\alias{fvsGetEventMonitorVariables}
\title{Get Event Monitor Variables}
\usage{
fvsGetEventMonitorVariables(vars)
}
\arguments{
\item{vars}{A scalar or vector of strings whereby each string names an FVS 
Event Monitor for which the value will be returned.}
}
\value{
A named vector of values. The values will be NA if they were not 
  defined.
}
\description{
Get Event Monitor Variables
}
