% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetRestartcode.R
\name{fvsGetRestartcode}
\alias{fvsGetRestartcode}
\title{Get Restart Code}
\usage{
fvsGetRestartcode()
}
\value{
The FVS restart codes (integer).
}
\description{
Provides an R interface to Fortran function fvsGetRestartcode.
}
