% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetSpeciesAttrs.R
\name{fvsGetSpeciesAttrs}
\alias{fvsGetSpeciesAttrs}
\title{Get Species Attributes}
\usage{
fvsGetSpeciesAttrs(vars)
}
\arguments{
\item{vars}{A vector of strings whereby each string names an attribute to be 
returned. The names can be one or more of the following:
`spccf`: CCF for each species, recomputed in FVS so setting will likely have 
no effect
`spsdi`: SDI maximums for each species
`spsiteindx`: Species site indices}
}
\value{
A dataframe of numeric values with one row for each species and a 
  column for each attribute.
}
\description{
Get Species Attributes
}
