% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetSpeciesCodes.R
\name{fvsGetSpeciesCodes}
\alias{fvsGetSpeciesCodes}
\title{Get Species Codes}
\usage{
fvsGetSpeciesCodes()
}
\value{
A matrix of character strings with one row for each species and three
  columns. The column names fvs, fia, plant correspond to the three kinds of 
  species codes used in FVS. The order is exactly as presented in the FVS 
  variant which means, for example, that row 1 corresponds to the internal 
  FVS species index number 1, row 2 for species 2 and so on.
}
\description{
Get Species Codes
}
