% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetStandIDs.R
\name{fvsGetStandIDs}
\alias{fvsGetStandIDs}
\title{Get Stand IDs}
\usage{
fvsGetStandIDs()
}
\value{
A named vector of three strings with names `standid`, `standcn`, and 
  `mgmtid.`
}
\description{
Get Stand IDs
}
