% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetSummary.R
\name{fvsGetSummary}
\alias{fvsGetSummary}
\title{Get Summary}
\usage{
fvsGetSummary()
}
\value{
A matrix of integer values with one row for each completed FVS cycle. 
The columns are named as documented in the FVS Summary Statistics.
}
\description{
Get Summary
}
