% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetSVSDims.R
\name{fvsGetSVSDims}
\alias{fvsGetSVSDims}
\title{Get SVS Dims}
\usage{
fvsGetSVSDims()
}
\value{
A named vector of integers as follows: 
  `nsvsobjs`, `ndeadobjs`, `ncwdobjs`, `mxsvsobjs`, `mxdeadobjs`, 
  `mxcwdobjs.`
}
\description{
Get SVS Dims
}
