% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetSVSObjectSet.R
\name{fvsGetSVSObjectSet}
\alias{fvsGetSVSObjectSet}
\title{Return the Stand Visualization System objects.}
\usage{
fvsGetSVSObjectSet()
}
\value{
a named list of trees, snags, and course woody debris (cwd). Each
item in the list is a data.frame of items.
}
\description{
Return the Stand Visualization System objects.
}
