% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetTreeAttrs.R
\name{fvsGetTreeAttrs}
\alias{fvsGetTreeAttrs}
\title{Get Tree Attrs}
\usage{
fvsGetTreeAttrs(vars)
}
\arguments{
\item{vars}{A vector of strings whereby each string names an attribute to be 
returned. The names can be one or more of the following:

see https://sourceforge.net/p/open-fvs/wiki/rFVS/#fvsgettreeattrsvars}
}
\value{
A dataframe of numeric values with one row for each tree and a column 
  for each attribute.
  
  The defect value is an 8-digit code comprised of four 2-digit code segments 
  strung together. Each 2-digit segment represents a different defect, as 
  described below.
  
  first and second digits: input merchantable cubic volume defect percent
  third and fourth digits: input board foot defect percent
  fifth and sixth digits: applied merchantable cubic volume defect percent
  seventh and eighth digits: applied board foot defect percent
}
\description{
Get Tree Attrs
}
