% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsLoad.R
\name{fvsLoad}
\alias{fvsLoad}
\title{Load FVS program}
\usage{
fvsLoad(fvsProgram, bin)
}
\arguments{
\item{fvsProgram}{Name of executable file for FVS variant.}

\item{bin}{File path to the directory containing FVS executables.}
}
\description{
Load FVS program
}
