% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsSetCmdLine.R
\name{fvsSetCmdLine}
\alias{fvsSetCmdLine}
\title{Load FVS program}
\usage{
fvsSetCmdLine(cl = NULL)
}
\arguments{
\item{cl}{A character string that contains the command line arguments. If
`NULL` (the default), the function attempts to fetch the arguments as
they may be specified on the command that starts R. This is done by
calling the R function `commandArgs(trailingOnly = TRUE)`, see R
documentation for details.}
}
\description{
Sets the command line to the values listed of the FVS Command
  Line and opens files for input as indicted by the values found. This
  routine can result in the FVS return code state to be reset.
}
