% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetEventMonitorVariables.R
\name{fvsSetEventMonitorVariables}
\alias{fvsSetEventMonitorVariables}
\title{Set Event Monitor Variables}
\usage{
fvsSetEventMonitorVariables(vars)
}
\arguments{
\item{vars}{A vector of named numeric values where the names are FVS Event 
Monitor variable names that will take on the corresponding values.}
}
\value{
A named vector of values. The values will be NA if they were not 
  defined.
}
\description{
This function sets the values of Event Monitor variables. 
  Generally, however, many of the variables are reset to by FVS when the 
  Event Monitor is called each cycle (before and after simulated harvests). 
  This is especially true of those automatic variables defined during the 
  first and second calls to the Event Monitor (consult FVS documentation). 
  Therefore setting the value of these variables may be in effective in 
  meeting useful goals. 
  
  One use of this function is to set user-defined variables (those 
  traditionally defined using the FVS Compute keyword). If a Event Monitor 
  variable that has the name given in the names attribute of the vars 
  argument is not found, one is created. If a user-defined variable has 
  already been defined, it is given a new value when this function is called.
}
