% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsSetTreeAttr.R
\name{fvsSetTreeAttrs}
\alias{fvsSetTreeAttrs}
\title{Set Tree Attrs}
\usage{
fvsSetTreeAttrs(vars)
}
\arguments{
\item{vars}{A named list of vectors (or a dataframe) where each vector (or 
column) is: of numeric type, has no NAs, has as many entries as there are 
defined species (variable maxspecies from fvsGetDims()).
Use the variable names listed for fvsGetSpeciesAttrs(vars). Vectors or 
columns that have other names are skipped.}
}
\value{
A scalar integer return code where 0 signals OK and 1 signals an 
  error).
}
\description{
Set Tree Attrs
}
