% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsSetupSummary.R
\name{fvsSetupSummary}
\alias{fvsSetupSummary}
\title{Setup Summary}
\usage{
fvsSetupSummary(asum)
}
\arguments{
\item{asum}{A matrix as returned by `fvsGetSummary()` or the summary table 
returned by `fvsCompositeSum(sumList)`.}
}
\value{
A matrix with additional columns and rows (if there were harvests 
  in the run).
}
\description{
Adds extra data to the summary statistics table returned by 
  function `fvsGetSummary()`. The extra data include new columns showing 
  total production of trees and volume plus new rows so that years with 
  removals have two data points, one for before and the after the harvest.
  
  Modifies the "summary" statistics so that it is ready to plot. This is done 
  by adding a row for post thin so that variables like "TPA" show the trace 
  over time with the removals being as sharp dip.
}
