% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetUnitConversion.R
\name{fvsUnitConversion}
\alias{fvsUnitConversion}
\title{Unit Conversion}
\usage{
fvsUnitConversion(name)
}
\arguments{
\item{name}{A case sensitive character string naming a conversion factor to 
be returned. See Return Unit Conversions for a list of valid codes.}
}
\value{
A scalar numeric value; NA if the name is not valid.
}
\description{
Unit Conversion
}
