% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsAddActivity.R
\name{fvsAddActivity}
\alias{fvsAddActivity}
\title{Add an an FVS Activity to the activity schedule}
\usage{
fvsAddActivity(year, activity, parms = NULL)
}
\arguments{
\item{year}{year the activity is to be scheduled}

\item{activity}{activity code as a number or character string}

\item{parms}{a numeric vector of parameters for the activity}
}
\value{
invisable return code with the value 0 if the activity was added and 1 if there was some error
}
\description{
Pass in an activity code and parameters to the FVS activity schedule.
}
