% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsAddTrees.R
\name{fvsAddTrees}
\alias{fvsAddTrees}
\title{Add new tree records to the simulation}
\usage{
fvsAddTrees(newtrees)
}
\arguments{
\item{newtrees}{newtrees a data.frame of new trees.}
}
\value{
invisable return code with the value 0 if the trees were added and
1 if there was some error
}
\description{
Pass a data.frame of new trees. The following columns must be specified:
"dbh","species","ht","cratio","plot", and "tpa".
}
