% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsCompositeSum.R
\name{fvsCompositeSum}
\alias{fvsCompositeSum}
\title{Build a composite summary statistics table}
\usage{
fvsCompositeSum(allsum)
}
\arguments{
\item{allsum}{a list of summary tables}
}
\value{
the composite table
}
\description{
A list of summary statistics tables are summarized into a composite (a
weighted average of values) for years that are common to all individual tables.
}
