% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetDims.R
\name{fvsGetDims}
\alias{fvsGetDims}
\title{Return the max dimensions of important FVS data storage.}
\usage{
fvsGetDims()
}
\value{
a named numeric vector.
}
\description{
Return the max dimensions of important FVS data storage.
}
