% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetEventMonitorVariables.R
\name{fvsGetEventMonitorVariables}
\alias{fvsGetEventMonitorVariables}
\title{Get Event Monitor Variables}
\usage{
fvsGetEventMonitorVariables(vars)
}
\arguments{
\item{vars}{a character vector of Event Monitor names.}
}
\value{
a named numeric vector of the variables, NA if the variable name does not exit
}
\description{
Get Event Monitor Variables
}
