% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetSpeciesAttrs.R
\name{fvsGetSpeciesAttrs}
\alias{fvsGetSpeciesAttrs}
\title{Return the values of species-specific data}
\usage{
fvsGetSpeciesAttrs(vars)
}
\arguments{
\item{vars}{a character vector of species-level attributes}
}
\value{
a data.frame where the columns are attributes and the rows are species.
}
\description{
Return the values of species-specific data
}
