% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetSpeciesCodes.R
\name{fvsGetSpeciesCodes}
\alias{fvsGetSpeciesCodes}
\title{Return a data.frame of species codes}
\usage{
fvsGetSpeciesCodes()
}
\value{
a data.frame with 1 row for each species and three character columns
of the "fvs", "fia", and "plant" codes.
}
\description{
Return a data.frame of species codes
}
