% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetStandIDs.R
\name{fvsGetStandIDs}
\alias{fvsGetStandIDs}
\title{Return stand identification codes}
\usage{
fvsGetStandIDs()
}
\value{
a names character vector containing the "standid", "standcn", "mgmtid", and "caseID"
}
\description{
Return stand identification codes
}
