% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetSummary.R
\name{fvsGetSummary}
\alias{fvsGetSummary}
\title{Return a data.frame of the summary statistics}
\usage{
fvsGetSummary()
}
\value{
summary statistics with one row per period
}
\description{
Return a data.frame of the summary statistics
}
