% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetSVSDims.R
\name{fvsGetSVSDims}
\alias{fvsGetSVSDims}
\title{Return the max dimensions of important Stand Visualization System data storage.}
\usage{
fvsGetSVSDims()
}
\value{
a named numeric vector.
}
\description{
Return the max dimensions of important Stand Visualization System data storage.
}
