% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetTreeAttrs.R
\name{fvsGetTreeAttrs}
\alias{fvsGetTreeAttrs}
\title{Get Attributes of Trees.}
\usage{
fvsGetTreeAttrs(vars)
}
\arguments{
\item{vars}{a character vector of tree attribute names.}
}
\value{
a data.frame with as many rows as there are tree records and a column
for each attribute.
}
\description{
Get Attributes of Trees.
}
