% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsInteractRun.R
\name{fvsInteractRun}
\alias{fvsInteractRun}
\title{Run FVS where FVS and R interact}
\usage{
fvsInteractRun(...)
}
\arguments{
\item{trace}{is true or false, when true informative messages are generated.}

\item{BeforeEM1}{R code to run at the stop point just before the first call to the Event Monitor}

\item{AfterEM1}{R code to run at the stop point just after the first call to the Event Monitor}

\item{BeforeEM2}{R code to run at the stop point just before the second call to the Event Monitor}

\item{AfterEM2}{R code to run at the stop point just after the second call to the Event Monitor}

\item{BeforeAdd}{R code to run at the stop point after growth and mortality has been computed, but prior to applying them}

\item{BeforeEstab}{R code to run at the stop point just before the Regeneration Establishment Model is called}

\item{SimEnd}{R code to run at the end of one stand simulation.}
}
\value{
a list of stands and years simulated with the call.
}
\description{
Pass one or more R-code blocks where each block is a named argument. FVS is run
up to the corresponding stop points and the code block is run. After the code block
is finished, FVS runs up to the next stop point. All of the code blocks are optional
in that any one or all can be used.
}
