% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsLoad.R
\name{fvsLoad}
\alias{fvsLoad}
\title{Load the desired FVS program as a shared library}
\usage{
fvsLoad(fvsProgram, bin = "../../trunk/bin")
}
\arguments{
\item{fvsProgram}{a character string of the name of the desired FVS shared library.}

\item{bin}{the name of the directory where the shared library is stored.}
}
\value{
the name that was loaded.
}
\description{
The FVS shared library is loaded if it is found. A side effect is that the name
that is loaded is stored in the R global environament as .FVSLOADEDLIBRARY
}
