% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsRun.R
\name{fvsRun}
\alias{fvsRun}
\title{Run FVS up to a designated stop point, year, or both.}
\usage{
fvsRun(stopPointCode = NA, stopPointYear = NA)
}
\arguments{
\item{stopPointCode}{is the integer value of the FVS stop point. If NA, no stop point is defined.}

\item{stopPointYear}{is the integer value of the year the stop should happen.}
}
\value{
the FVS return code where -1 indicates that FVS has not been started,
0 indicates that FVS is in good running state, 1 indicates that FVS has detected
an error of some kind and should not be used until reset by specifying new input and
2 indicates that FVS has finished processing all the stands; new input can be specified..
}
\description{
Run FVS up to a designated stop point, year, or both.
}
