% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsSetCmdLine.R
\name{fvsSetCmdLine}
\alias{fvsSetCmdLine}
\title{Sets the command line}
\usage{
fvsSetCmdLine(cl = NULL)
}
\arguments{
\item{cl}{character string of the FVS command line.}
}
\value{
the FVS return code or NULL if there is no command line specified.
}
\description{
Sets the command line
}
