% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetEventMonitorVariables.R
\name{fvsSetEventMonitorVariables}
\alias{fvsSetEventMonitorVariables}
\title{Set Event Monitor Variables}
\usage{
fvsSetEventMonitorVariables(vars)
}
\arguments{
\item{vars}{a named numeric vector of the variables and corresponding values.}
}
\value{
a named numeric vector of the variables, NA if the variable name does not exit
}
\description{
Set Event Monitor Variables
}
