% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsSetSpeciesAttr.R
\name{fvsSetSpeciesAttrs}
\alias{fvsSetSpeciesAttrs}
\title{Set the values of species-specific data}
\usage{
fvsSetSpeciesAttrs(vars)
}
\arguments{
\item{vars}{a named list of numeric vectors where the name corresponds to an attribute and
the vector contains values for each species (in order).}
}
\value{
scalar integer 0 signals OK and 1 signals an error.
}
\description{
Set the values of species-specific data
}
