% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsSetTreeAttr.R
\name{fvsSetTreeAttrs}
\alias{fvsSetTreeAttrs}
\title{Set Attributes of Trees}
\usage{
fvsSetTreeAttrs(vars)
}
\arguments{
\item{vars}{a named list of numeric vectors where the names are attributes
and the vector contains values for each tree (in order).}
}
\value{
scalar integer 0 signals OK and 1 signals an error.
}
\description{
Set Attributes of Trees
}
