% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsSetupSummary.R
\name{fvsSetupSummary}
\alias{fvsSetupSummary}
\title{Modifies the "summary" statistics so that it is ready to plot}
\usage{
fvsSetupSummary(asum)
}
\arguments{
\item{asum}{as returned from \code{\link[=fvsGetSummary]{fvsGetSummary()}}}
}
\value{
A reorganized version of \code{\link{asum}}.
}
\description{
The modification is done by adding a row for post thin statistics so that
variables like "TPA" show the trace over time with the removals being
shown as a vertical drop to the post thin level.
}
\examples{
fvsSetupSummary(fvsGetSummary())
}
