% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvsGetUnitConversion.R
\name{fvsUnitConversion}
\alias{fvsUnitConversion}
\title{Get FVS Units Conversion Factor}
\usage{
fvsUnitConversion(name)
}
\arguments{
\item{name}{a character string name of the desired conversion factor.}
}
\value{
the numeric value, or NA if FVS does not contain the desired factor.
}
\description{
Get FVS Units Conversion Factor
}
